修改商品活动

+++ src/main/java/com/neu/shop/controller/front/OrderController.java
@@ -96,7 +96,7 @@ public class OrderController {
 
     @RequestMapping("/orderFinish")
     @ResponseBody
-    public Msg orderFinish(Integer oldPrice, Integer newPrice, Boolean isPay, Integer addressid,HttpSession session) {
+    public Msg orderFinish(Integer oldPrice, Float newPrice, Boolean isPay, Integer addressid,HttpSession session) {
         User user = (User) session.getAttribute("user");
 
         //获取订单信息
